+++ TipPlace/TipPlace/CategoryBoard/RecommendList.swift
@@ -26,8 +26,8 @@ struct RecommendList: View {
     }
 }
 
-//struct RecommendList_Previews: PreviewProvider {
-//    static var previews: some View {
-//        RecommendList(categoryEnum: Category.livingAlone)
-//    }
-//}
+struct RecommendList_Previews: PreviewProvider {
+    static var previews: some View {
+        RecommendList(categoryEnum: Category.livingAlone)
+    }
+}
