+++ TipPlace/TipPlace/ModelData/ListMock.swift
@@ -1,0 +1,91 @@
+//
+//  MockData.swift
+//  TipPlace
+//
+//  Created by 리아 on 2022/04/10.
+//
+
+import Foundation
+
+struct ListMock {
+    static var boardPosts: [BoardPost] = [BoardPostMock.boardPost1,
+                                          BoardPostMock.boardPost2,
+                                          BoardPostMock.boardPost3,
+                                          BoardPostMock.boardPost4,
+                                          BoardPostMock.boardPost5,
+                                          BoardPostMock.boardPost53,
+                                          BoardPostMock.boardPost54,
+                                          BoardPostMock.boardPost55,
+                                          BoardPostMock.boardPost56,
+                                          BoardPostMock.boardPost57]
+    static var authors: [Author] = [AuthorMock.author1,
+                                    AuthorMock.author2,
+                                    AuthorMock.author3,
+                                    AuthorMock.author4,
+                                    AuthorMock.author5,
+                                    AuthorMock.author6,
+                                    AuthorMock.author7,
+                                    AuthorMock.author8,
+                                    AuthorMock.author9,
+                                    AuthorMock.author10,
+                                    AuthorMock.author53,
+                                    AuthorMock.author54,
+                                    AuthorMock.author55,
+                                    AuthorMock.author56,
+                                    AuthorMock.author57]
+    static var userInfos: [UserInfo] = [UserInfoMock.userInfo1,
+                                        UserInfoMock.userInfo2,
+                                        UserInfoMock.userInfo3,
+                                        UserInfoMock.userInfo4,
+                                        UserInfoMock.userInfo5,
+                                        UserInfoMock.userInfo6,
+                                        UserInfoMock.userInfo7,
+                                        UserInfoMock.userInfo8,
+                                        UserInfoMock.userInfo9,
+                                        UserInfoMock.userInfo10,
+                                        UserInfoMock.userInfo53,
+                                        UserInfoMock.userInfo54,
+                                        UserInfoMock.userInfo55,
+                                        UserInfoMock.userInfo56,
+                                        UserInfoMock.userInfo57]
+    static var detailPosts: [DetailPostModel] = [DetailPostMock.detailedPost1,
+                                                 DetailPostMock.detailedPost2,
+                                                 DetailPostMock.detailedPost3,
+                                                 DetailPostMock.detailedPost4,
+                                                 DetailPostMock.detailedPost5,
+                                                 DetailPostMock.DetailPostMock53,
+                                                 DetailPostMock.DetailPostMock54,
+                                                 DetailPostMock.DetailPostMock55,
+                                                 DetailPostMock.DetailPostMock56,
+                                                 DetailPostMock.DetailPostMock57]
+    static var comments: [CommentModel] = [CommentModelMock.CommentModelMock1,
+                                           CommentModelMock.CommentModelMock2,
+                                           CommentModelMock.CommentModelMock3,
+                                           CommentModelMock.CommentModelMock4,
+                                           CommentModelMock.CommentModelMock5,
+                                           CommentModelMock.CommentModelMock6,
+                                           CommentModelMock.CommentModelMock7,
+                                           CommentModelMock.CommentModelMock8,
+                                           CommentModelMock.CommentModelMock9,
+                                           CommentModelMock.CommentModelMock10,
+                                           CommentModelMock.CommentModelMock11,
+                                           CommentModelMock.CommentModelMock12,
+                                           CommentModelMock.CommentModelMock13,
+                                           CommentModelMock.CommentModelMock14,
+                                           CommentModelMock.CommentModelMock15,
+                                           CommentModelMock.CommentModelMock16,
+                                           CommentModelMock.CommentModelMock53,
+                                           CommentModelMock.CommentModelMock54,
+                                           CommentModelMock.CommentModelMock55,
+                                           CommentModelMock.CommentModelMock56,
+                                           CommentModelMock.CommentModelMock57,
+                                           CommentModelMock.CommentModelMock58,
+                                           CommentModelMock.CommentModelMock59,
+                                           CommentModelMock.CommentModelMock60,
+                                           CommentModelMock.CommentModelMock61,
+                                           CommentModelMock.CommentModelMock62,
+                                           CommentModelMock.CommentModelMock63,
+                                           CommentModelMock.CommentModelMock64,
+                                           CommentModelMock.CommentModelMock65,
+                                           CommentModelMock.CommentModelMock66]
+}
