+++ TipPlace/TipPlace/Utility/loadPosts.swift
@@ -15,7 +15,8 @@ func filterPost(with allPosts: [BoardPost], by selected: String) -> [BoardPost] 
         return allPosts
     } else {
         return allPosts
-            .filter({ $0.tag.contains(selected)})
+            .filter({
+                $0.tag.contains(selected)})
 //        옵셔널 -> 빈 배열 수정
             .sorted(by: { $0.usefulCount > $1.usefulCount })
     }
